fix: contestants not getting assigned rating
Problem:
The map with rating as key resulted in only one rating for every
rank. This meant that every contestant except 1 from every rank had
rating 0.

Solution:
Create map with handle as key and a RatingChange as value. Then choose
the highest OldRating of the partymembers for each contestant.

This is pretty much the same issue as #57