refactor: remove nginx from project
It is no longer needed as Browsersync does everything nginx did for
development porposes. For deploying the site on a dedicated server nginx
should still be used, but that config does not need to live in this repo
as the same nginx instance would be responsible for everything running
on the server.