feat: use nginx for https and http/2
This allows for request cancellation, so that when the client cancels we
can cancel the connected requests on the server. The current setup is
that we connect to nginx on :443, which proxies to Browsersync on :3000,
and proxies /api requests to the server on :8080. This way HTTP/2 works
for the backend requests.

Also fixed the Store.GetContestResults function to not try the API if
the DB query failed due to cancelled context.
(I know I removed nginx a couple commits ago, dont worry about it).